Don't allow unaligned reads or writes.
This patch is for emscripten and platforms that do not support unaligned
reads and writes (such as ARM). memcpy is used to copy out of char
buffers and into aligned memory. uint32_t and uint64_t pointers are no
longer used in Tny_swapBytes32 and Tny_swapBytes64, respectively. The
pointers have been replaced by char pointers.